* More work on the menus * Add a menu bar (need some theming) * Add a status bar (the resize grip doesn't work yet) * Clean up etk_test * Some minor fixes and new things
git-svn-id: http://svn.enlightenment.org/svn/e/trunk@18547 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33

+++ e17/libs/etk/src/lib/etk_signal.c
@@ -60,7 +60,7 @@ void etk_signal_shutdown()
  * @param accum_data the value to pass to the accumulator
  * @return Returns the new signal, or NULL on failure
  */
-Etk_Signal *etk_signal_new(const char *signal_name, Etk_Type *object_type, int default_handler_offset, Etk_Marshaller marshaller, Etk_Accumulator accumulator, void *accum_data)
+Etk_Signal *etk_signal_new(const char *signal_name, Etk_Type *object_type, long default_handler_offset, Etk_Marshaller marshaller, Etk_Accumulator accumulator, void *accum_data)
 {
    Etk_Signal *new_signal;
